Optimizing RNG

+++ includes/randomc.hpp
@@ -11,8 +11,8 @@ class IRand
     public:
         virtual ~IRand();
 
-        virtual long randomInit(long) = 0; //re-seed generator
-        virtual long intRandom(long, long) = 0; //random integer
+        virtual long randomInit(unsigned int) = 0; //re-seed generator
+        virtual unsigned int intRandom(long, long) = 0; //random integer
         virtual long int getSeed() = 0;
 };
 
@@ -23,14 +23,14 @@ class simpleRandom : public IRand
         simpleRandom();
         ~simpleRandom();
 
-        inline long randomInit(long int);
-        inline long intRandom(long, long);
+        inline long randomInit(unsigned int);
+        inline unsigned int intRandom(long, long);
         inline long int getSeed() { return seed; }
 
     private:
         inline unsigned int random();
 
-        unsigned long next;
+        unsigned long int next;
         unsigned long seed;
 };
 
@@ -39,8 +39,8 @@ class nullRNG : public IRand
     public:
         nullRNG() { logger=NULL; }
         ~nullRNG() {}
-        inline long randomInit(long int value) { log(); return 0; }
-        inline long intRandom(long value1, long value2) { log(); return 0; }
+        inline long randomInit(unsigned int value) { log(); return 0; }
+        inline unsigned int intRandom(long value1, long value2) { log(); return 0; }
         inline long getSeed() { log(); return 0; }
 
     private:
